Added an AI player that guesses based on rank
Added a simple reflex agent that looks at the rank of the card, see how
far above/below the midpoint in the sequence of ranks, and then selects
"higher" or "lower" based on which half it is in. This is my typical strategy
when I play this game.

Soon I'll add card counting to see if I can increase the average number of
turns past 3.3 turns

NOTE: Aces are low for a simpler implementation here that will be consistent
with both playing cards and ByteCards.